Refactor imports to base numerous module
