+++ examples/export_example.py
@@ -5,8 +5,7 @@ from enum import Enum
 from numerous.engine.model import Model
 from numerous.engine.simulation import Simulation
 
-from numerous.engine.system import Subsystem, ConnectorItem, Item, ConnectorTwoWay
-from numerous import EquationBase, Equation
+from numerous.engine.system import Subsystem, Item
 from numerous.engine.simulation.solvers.base_solver import SolverType
 from tests.test_equations import TestEq_ground, Test_Eq, TestEq_input
