+++ numerous/engine/model/graph_representation/mappings_graph.py
@@ -1,6 +1,6 @@
+from numerous.engine import VariableType, SetOfVariables
 from numerous.engine.model.graph_representation.utils import EdgeType
 from numerous.engine.model.utils import NodeTypes
-from numerous import VariableType, SetOfVariables, Variable
 from numerous.utils.string_utils import d_u
 
 from .graph import Graph, Node, Edge
