+++ tests/test_binding.py
@@ -1,7 +1,7 @@
 import pytest
 
+from numerous.engine import VariableDescription, VariableType
 from numerous.engine.system import ConnectorTwoWay, Item
-from numerous import VariableDescription, VariableType
