+++ tests/test_item_subsystem.py
@@ -1,7 +1,7 @@
 import pytest
 
+from numerous.engine import VariableDescription, VariableType
 from numerous.engine.system import Connector, Item
-from numerous import VariableDescription, VariableType
