+++ tests/test_not_used_variable.py
@@ -1,11 +1,11 @@
 import pytest
 from pytest import approx
 
-from numerous import EquationBase, Equation
 from numerous.engine.model import Model
 from numerous.engine.simulation import Simulation
 from numerous.engine.simulation.solvers.base_solver import solver_types
 from numerous.engine.system import Subsystem, Item
+from numerous.multiphysics import EquationBase, Equation
 import numpy as np
