+++ tests/test_var.py
@@ -1,7 +1,6 @@
 import pytest
 
-from numerous.engine.variables import _VariableFactory, OverloadAction, VariableType, Variable
-from numerous import VariableDescription
+from numerous.engine.variables import _VariableFactory, OverloadAction, VariableType, Variable, VariableDescription
 from tests.test_equations import *
